style: move segmented_log tests to in_mem::commit_log::segmented_log
